Revert "Restore round off-ed values and remove unnecessary code"
This reverts commit cb830f62405ee09a44227064544874002c4520ac.

+++ src/sense360/UserStayInfo.java
@@ -2,8 +2,8 @@ package sense360;
 
 public class UserStayInfo {
 	Point point;
-	int totalHours;
-	int nightHours;
+	long totalHours;
+	long nightHours;
 
 	public UserStayInfo(Point point) {
 		this.point = point;
